Nyt koodi on vähän järkevämpää ja toimii (kuitenkin tee paremmaksi myöhemmin)
